abstracted init and update version logic into core
