+++ pkg/commands/init.go
@@ -1,13 +1,11 @@
 package commands
 
 import (
-	"io"
-	"os"
-
-	"github.com/AlecAivazis/survey/v2"
-	"github.com/spf13/cobra"
+	"fmt"
 
 	"github.com/restechnica/semverbot/pkg/cli"
+	"github.com/restechnica/semverbot/pkg/core"
+	"github.com/spf13/cobra"
 )
 
 // NewInitCommand creates a new init command.
@@ -16,7 +14,7 @@ func NewInitCommand() *cobra.Command {
 	var command = &cobra.Command{
 		Use:   "init",
 		RunE:  InitCommandRunE,
-		Short: "Creates a default .semverbot.toml config",
+		Short: fmt.Sprintf(`Creates a default "%s" config`, cli.DefaultConfigFilePath),
 	}
 
 	return command
@@ -25,32 +23,10 @@ func NewInitCommand() *cobra.Command {
 // InitCommandRunE runs the init command.
 // Returns an error if the command failed.
 func InitCommandRunE(cmd *cobra.Command, args []string) (err error) {
-	var file *os.File
-
-	if _, err = os.Stat(cli.DefaultConfigFilePath); !os.IsNotExist(err) {
-		var prompt = &survey.Confirm{
-			Message: "Do you wish to overwrite your current config?",
-		}
-
-		var isOk = false
-
-		if err = survey.AskOne(prompt, &isOk); err != nil {
-			return err
-		}
-
-		if !isOk {
-			return
-		}
+	var options = &core.InitOptions{
+		Config:         cli.DefaultConfig,
+		ConfigFilePath: cli.DefaultConfigFilePath,
 	}
 
-	if file, err = os.Create(cli.DefaultConfigFilePath); err != nil {
-		return err
-	}
-
-	if _, err = io.WriteString(file, cli.DefaultConfig); err != nil {
-		_ = file.Close()
-		return err
-	}
-
-	return file.Close()
+	return core.Init(options)
 }
