+++ pkg/commands/update-version.go
@@ -1,9 +1,8 @@
 package commands
 
 import (
-	"fmt"
+	"github.com/restechnica/semverbot/pkg/core"
 
-	"github.com/restechnica/semverbot/pkg/api"
 	"github.com/spf13/cobra"
 )
 
@@ -18,20 +17,8 @@ func NewUpdateVersionCommand() *cobra.Command {
 	return command
 }
 
-// UpdateVersionCommandRunE runs before the commands runs.
+// UpdateVersionCommandRunE runs the commands.
 // Returns an error if it fails.
 func UpdateVersionCommandRunE(cmd *cobra.Command, args []string) (err error) {
-	var gitAPI = api.NewGitAPI()
-
-	if err = gitAPI.FetchUnshallow(); err != nil {
-		fmt.Println("something went wrong while fetching from git, attempting to fetch tags anyway")
-	}
-
-	if err = gitAPI.FetchTags(); err != nil {
-		fmt.Println("something went wrong while updating the version")
-	} else {
-		fmt.Println("successfully fetched the latest git tags")
-	}
-
-	return err
+	return core.UpdateVersion()
 }
